ai-in-action/2-diagram-generation-agents/requirements.txt
colorama
python-dotenv
typing
smolagents
litellm
mermaid-py
